feat(genealogy): add DrawerAnimals component and integrate into ReproductionCreateModal
This commit introduces a new `DrawerAnimals` component to facilitate animal selection in the genealogy module. The component is integrated into the `ReproductionCreateModal` to allow users to select mother and father animals from a drawer interface. This improves the user experience by providing a more intuitive way to select animals for reproduction records.